⚡  Updated distanceAway to double ⚡  changed "about" to "~" for distance

+++ lib/models/api/transito/nearby_bus_stops.dart
@@ -21,7 +21,7 @@ class NearbyBusStopsApiResponse {
 @JsonSerializable(explicitToJson: true, createToJson: false)
 class NearbyBusStop {
   BusStopInfo busStop;
-  int distanceAway;
+  double distanceAway;
 
   NearbyBusStop({
     required this.busStop,
